Conflicts:
	palladian-core/src/integration-test/java/ws/palladian/classification/featureselection/FeatureRankerIT.java
	palladian-core/src/main/java/ws/palladian/classification/Instance.java
	palladian-core/src/main/java/ws/palladian/classification/featureselection/AbstractFeatureRanker.java
	palladian-core/src/main/java/ws/palladian/classification/featureselection/BackwardFeatureElimination.java
	palladian-core/src/main/java/ws/palladian/classification/utils/ClassificationUtils.java
	palladian-core/src/main/java/ws/palladian/extraction/location/FeatureBasedDisambiguation.java
	palladian-core/src/main/java/ws/palladian/extraction/pos/BasePosTagger.java
	palladian-core/src/main/java/ws/palladian/extraction/token/BaseTokenizer.java
	palladian-core/src/main/java/ws/palladian/extraction/token/LingPipeTokenizer.java
	palladian-core/src/main/java/ws/palladian/extraction/token/OpenNlpTokenizer.java
	palladian-core/src/main/java/ws/palladian/extraction/token/RegExTokenizer.java
	palladian-core/src/test/java/ws/palladian/classification/dt/DecisionTreeClassifierTest.java
	palladian-core/src/test/java/ws/palladian/extraction/pos/LingPipePosTaggerTest.java
	palladian-external-code/src/test/java/ws/palladian/classification/WekaPredictorTest.java
	palladian-processing-api/src/main/java/ws/palladian/processing/features/FeatureVector.java

Former-commit-id: 734429d1f12c74b8de6e3fea3491363f539a8501

+++ palladian-retrieval/src/main/java/ws/palladian/retrieval/search/web/WebSearcher.java
@@ -86,7 +86,7 @@ public abstract class WebSearcher<R extends WebResult> implements Searcher<R> {
     public abstract List<R> search(String query, int resultCount, Language language) throws SearcherException;
 
     @Override
-    public final int getTotalResultCount(String query) throws SearcherException {
+    public final long getTotalResultCount(String query) throws SearcherException {
         return getTotalResultCount(query, DEFAULT_SEARCHER_LANGUAGE);
     }
 
@@ -100,7 +100,7 @@ public abstract class WebSearcher<R extends WebResult> implements Searcher<R> {
      * @return
      * @throws SearcherException In case the search fails.
      */
-    public int getTotalResultCount(String query, Language language) throws SearcherException {
+    public long getTotalResultCount(String query, Language language) throws SearcherException {
         throw new SearcherException("Obtaining the total number of results is not supported or implemented by "
                 + getName() + ".");
     }
